requirements.txt
fastapi
uvicorn
python-multipart
timm
torch
torchvision
Pillow
numpy
requests
opencv-python